:recycle: api: move usecase endAssessmentBySupervisor to src
