+++ api/src/certification/course/domain/usecases/end-assessment-by-supervisor.js
